+++ changelog.md
@@ -63,6 +63,7 @@
 - Fixed ice cubes freezing owners & teammates
 - Fixed pipes not connecting properly in structures during world gen
 - Fixed super mushrooms not generating in bonus chests
+- Fixed reward particles not rendering with minimal particles enabled
 
 ## Update 1.9.3
 - Added Sable block tags for Create Aeronautics compat